+++ redrock/lesson5/utils/utils.go
@@ -29,7 +29,7 @@ func Jiami(Password string) string {
 
 type CustomClaims struct {
 	//用户ID，简称UID，原神...
-	UserID uint64 `json:"uid"`
+	UserID uint `json:"uid"`
 	//用户角色，可以精确的推送小广告？
 	Role string `json:"role"`
 	//记录访问令牌还是刷新令牌
@@ -41,7 +41,7 @@ type CustomClaims struct {
 	jwt.RegisteredClaims
 }
 
-func GenerateTokens(userID uint64, role string) (accessToken string, refreshToken string, err error) {
+func GenerateTokens(userID uint, role string) (accessToken string, refreshToken string, err error) {
 	now := time.Now()
 	accessClaims := CustomClaims{
 		UserID: userID,
